A major reorganization:
* Changed the groupId from org.l2x6.maven.srcdeps to org.l2x6.srcdeps
because we have the ambition to stop being Maven specific now.
Eventually the srcdeps-maven-local-repository that indeed is Maven
specific may move to a separate repo in the future.
* Added srcdeps-maven-local-repository that contains an implementation
of org.eclipse.aether.repository.LocalRepositoryManager which decorates
any LocalRepositoryManager available in the class path by the ability to
build the requested artifact from sources.
* Added srcdeps-config-yaml